requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
ta-lib==0.4.28
ccxt==4.1.59
pytest==7.4.3
black==23.12.1
